+++ lib/plugins/aws/metrics.js
@@ -4,7 +4,7 @@ const BbPromise = require('bluebird');
 const chalk = require('chalk');
 const _ = require('lodash');
 const dayjs = require('dayjs');
-const validate = require('../lib/validate');
+const validate = require('./lib/validate');
 
 const LocalizedFormat = require('dayjs/plugin/localizedFormat');
