+++ lib/plugins/index.js
@@ -28,7 +28,7 @@ module.exports = [
   require('./aws/invoke.js'),
   require('./aws/info/index.js'),
   require('./aws/logs.js'),
-  require('./aws/metrics/awsMetrics.js'),
+  require('./aws/metrics.js'),
   require('./aws/remove/index.js'),
   require('./aws/rollback/index.js'),
   require('./aws/rollbackFunction/index.js'),
